fix(paseo): bind hook ownership and split inventory contracts
Validate Grok paths against the active context across generated launch forms. Preserve the inventory API through focused modules. Keep all current main security gates.
